% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{latest_release}
\alias{latest_release}
\title{The most recent bundled PSGC release}
\usage{
latest_release()
}
\value{
A single character string naming the latest available release.
}
\description{
The most recent bundled PSGC release
}
\examples{
latest_release()
}
