% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{list_releases}
\alias{list_releases}
\title{List available PSGC releases}
\usage{
list_releases()
}
\value{
A character vector of release names in chronological order.
}
\description{
List available PSGC releases
}
\examples{
list_releases()
}
